Parameterize --os-variant flag to virt-install
When deploying with other virt-install versions, it's possible that not
all os-variants might exist on the system. This change updates the
hardcoded linux2022 os-variant to be dynamic, leaving the default set to
the previous version for backwards compatibility.

Also changed is an update to the rhel-9-large configuration template
with the new OS_VARIANT parameter and setting the value to 'rhl9'. All
available os-variants available to the system can be found by running
'osinfo-query os'.

Signed-off-by: Leif Madsen <lmadsen@redhat.com>