Add DISK_EXPAND_PART parameter
Add the DISK_EXPAND_PART parameter to the init script in order to allow
a configuration to pass which device partition to expand during
deployment. By default the device partition is /dev/sda1 which on at
least RHEL 9.0, is the BIOS boot partition.

The general desire would be to expand the root partition, which by
default is limited to 10G. An administrator is now able to provide the
DISK_EXPAND_PART=/dev/sda4 (for example) within the configuration for
the deployment, to apply the virt-install --expand parameter to the
appropriate disk partition.

Updated the rhel-9-large.conf after verification of this required change
to that template. Other templates may also require this, but were not
tested.